backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Noor E Shams <shamsniloy75@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.7"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.7.0"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
